model/FACE_cv/requirements.txt
numpy==1.24.3
opencv-python==4.8.0.74
torch==2.1.0
torchvision==0.16.0
facenet-pytorch==2.5.3
mysql-connector-python==8.1.0
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.24.0
pillow==10.1.0 